refactor akka common
