+++ build.gradle
@@ -39,6 +39,7 @@ buildscript {
 }
 
 ext.versions = [
+        akka     : "2.6.15",
         commons  : "3.5",
         flink    : "1.14.2",
         hadoop   : "2.7.7",
@@ -47,12 +48,21 @@ ext.versions = [
         junit5   : "5.4.+",
         mockito  : "2.0.+",
         mockito3 : "3.+",
+        scala    : "2.12",
         spectator: "1.3.+",
         slf4j    : "1.7.0",
         vavr     : "0.9.2",
 ]
 
 ext.libraries = [
+        akkaActor      : "com.typesafe.akka:akka-actor_${versions.scala}:${versions.akka}",
+        akkaHttp       : "com.typesafe.akka:akka-http_${versions.scala}:${versions.akka}",
+        akkaHttpJackson: "com.typesafe.akka:akka-http-jackson_${versions.scala}:${versions.akka}",
+        akkaHttpCaching: "com.typesafe.akka:akka-http-caching_${versions.scala}:${versions.akka}",
+        akkaStream     : "com.typesafe.akka:akka-stream_${versions.scala}:${versions.akka}",
+        akkaSlf4j      : "com.typesafe.akka:akka-slf4j_${versions.scala}:${versions.akka}",
+        akkaTest       : "com.typesafe.akka:akka-testkit_${versions.scala}:${versions.akka}",
+        akkaHttpTest   : "com.typesafe.akka:akka-http-testkit_${versions.scala}:${versions.akka}",
         asyncHttpClient: "org.asynchttpclient:async-http-client:2.12.3",
         commonsIo      : "commons-io:commons-io:2.11.0",
         commonsLang3   : "org.apache.commons:commons-lang3:${versions.commons}",
