+++ mantis-common-akka/build.gradle
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2024 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+dependencies {
+    api project(":mantis-common")
+    // spectatorApi should be packaged at entry point level to avoid version conflicts.
+    compileOnly libraries.spectatorApi
+
+    api libraries.akkaActor
+
+    testImplementation libraries.akkaTest
+    testImplementation libraries.spectatorApi
+    testImplementation libraries.junit4
+    testImplementation libraries.mockitoCore
+}
