+++ mantis-common-akka/src/main/java/io/mantisrx/common/akka/ActorSystemMetrics.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.mantisrx.master.akka;
+package io.mantisrx.common.akka;
 
 import io.mantisrx.common.metrics.Counter;
 import io.mantisrx.common.metrics.Metrics;
