+++ mantis-common-akka/src/main/java/io/mantisrx/common/akka/DeadLetterActor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.mantisrx.master;
+package io.mantisrx.common.akka;
 
 import akka.actor.AbstractActor;
 import akka.actor.DeadLetter;
