+++ mantis-common-akka/src/main/java/io/mantisrx/common/akka/MantisActorSupervisorStrategy.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.mantisrx.master.akka;
+package io.mantisrx.common.akka;
 
 import akka.actor.ActorInitializationException;
 import akka.actor.ActorKilledException;
