+++ mantis-common-akka/src/main/java/io/mantisrx/common/akka/MeteredMessageQueue.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.mantisrx.master.akka;
+package io.mantisrx.common.akka;
 
 import akka.actor.ActorRef;
 import akka.dispatch.Envelope;
