+++ mantis-control-plane/mantis-control-plane-server/src/main/resources/application.conf
@@ -101,7 +101,7 @@ akka {
     # The guardian "/user" will use this class to obtain its supervisorStrategy.
     # It needs to be a subclass of akka.actor.SupervisorStrategyConfigurator.
     # In addition to the default there is akka.actor.StoppingSupervisorStrategy.
-    guardian-supervisor-strategy = "io.mantisrx.master.akka.MantisActorSupervisorStrategy"
+    guardian-supervisor-strategy = "io.mantisrx.common.akka.MantisActorSupervisorStrategy"
 
     # Timeout for ActorSystem.actorOf
     creation-timeout = 20s
@@ -625,7 +625,7 @@ akka {
       # constructor with
       # (akka.actor.ActorSystem.Settings, com.typesafe.config.Config) parameters.
       mailbox-type = "akka.dispatch.UnboundedMailbox"
-      # mailbox-type = "io.mantisrx.master.akka.UnboundedMeteredMailbox"
+      # mailbox-type = "io.mantisrx.common.akka.UnboundedMeteredMailbox"
 
       # If the mailbox is bounded then it uses this setting to determine its
       # capacity. The provided value must be positive.
@@ -651,7 +651,7 @@ akka {
       # constructor with
       # (akka.actor.ActorSystem.Settings, com.typesafe.config.Config) parameters.
       # mailbox-type = "akka.dispatch.UnboundedMailbox"
-      mailbox-type = "io.mantisrx.master.akka.UnboundedMeteredMailbox"
+      mailbox-type = "io.mantisrx.common.akka.UnboundedMeteredMailbox"
 
       # If the mailbox is bounded then it uses this setting to determine its
       # capacity. The provided value must be positive.
