+++ mantis-runtime-executor/src/main/java/io/mantisrx/server/worker/jobmaster/JobMasterServiceV2.java
@@ -1,0 +1,4 @@
+package io.mantisrx.server.worker.jobmaster;
+
+public class JobMasterServiceV2 {
+}
