+++ settings.gradle
@@ -22,6 +22,7 @@ rootProject.name = 'mantis'
 
 include 'mantis-client'
 include 'mantis-common'
+include 'mantis-common-akka'
 include 'mantis-common-serde'
 
 include 'mantis-connectors:mantis-connector-iceberg'
